pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-prompt-injection"
version = "0.1.0"
description = "Indirect Prompt Injection in RAG-Heavy Enterprise Search Copilot"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    # RAG Pipeline
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-llms-ollama>=0.3.0",
    "llama-index-vector-stores-chroma>=0.2.0",
    # Vector DB
    "chromadb>=0.5.0",
    # Reranker
    "sentence-transformers>=3.0.0",
    # ML Detector
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "datasets>=2.18.0",
    # Data Generation
    "faker>=24.0.0",
    "pyyaml>=6.0",
    # Config
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    # Evaluation & Visualization
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
